Restructure JavaScript code in multiple files
