+++ src/main/webapp/index.html
@@ -5,8 +5,8 @@
     <link rel="stylesheet" href="style.css">
     <link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Assistant">
     <meta name="google-signin-client_id" content="">
-    <script src="script.js"></script>
-    <script async defer src="https://apis.google.com/js/api.js" onload="loadClient()">
+    <script src="script.js" type="module"></script>
+    <script async defer src="https://apis.google.com/js/api.js">
     </script>
   </head>
   <body>
